spm: Add support for SPM
