+++ Sources/Constraid/CenterWithinMargins.swift
@@ -1,6 +1,7 @@
 // We don't conditional import AppKit like normal here because AppKit Autolayout doesn't support
 // the margin attributes that UIKit does. And of course this file isn't included in the MacOS
 // build target.
+#if os(iOS)
 import UIKit
 
 /**
@@ -136,3 +137,5 @@ public func center(_ itemA: Constraid.View, withinMarginsOf itemB: Any?,
         return constraints
     }
 }
+
+#endif
