+++ Sources/Constraid/LimitByEdge.swift
