+++ Tests/ConstraidTests/FlushWithEdgesTests.swift
@@ -1,72 +1,52 @@
 import XCTest
 import Constraid
 
-class ExpandFromEdgeTests: XCTestCase {
-    func testExpandFromLeadingEdgeOf() {
+#if os(iOS)
+import UIKit
+
+class FlushWithEdgesTests: XCTestCase {
+
+    func testFlushWithLeadingEdgeOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-
-        let constraints = Constraid.expand(viewOne, fromLeadingEdgeOf: viewTwo, times: 2.0, offsetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        let constraints = Constraid.flush(viewOne, withLeadingEdgeOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
         constraints.activate()
+
         let constraintOne = viewOne.constraints.first!
 
         XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.lessThanOrEqual)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.constant, 10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
 
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testExpandByTrailingEdgeOf() {
+    func testFlushWithTrailingEdgeOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-
-        let constraints = Constraid.expand(viewOne, fromTrailingEdgeOf: viewTwo, times: 2.0, offsetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        let constraints = Constraid.flush(viewOne, withTrailingEdgeOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
         constraints.activate()
+
         let constraintOne = viewOne.constraints.first!
 
         XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.greaterThanOrEqual)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintOne.constant, 10.0)
-        XCTAssertEqual(constraintOne.multiplier, 2.0)
-        XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
-
-        XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
-    }
-
-    func testExpandFromTopEdgeOf() {
-        let viewOne = UIView()
-        let viewTwo = UIView()
-
-        viewOne.addSubview(viewTwo)
-
-        let constraints = Constraid.expand(viewOne, fromTopEdgeOf: viewTwo, times: 2.0, offsetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
-        constraints.activate()
-        let constraintOne = viewOne.constraints.first!
-
-        XCTAssertEqual(constraints, viewOne.constraints)
-        XCTAssertEqual(constraintOne.isActive, true)
-        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
-        XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.top)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.lessThanOrEqual)
-        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.top)
         XCTAssertEqual(constraintOne.constant, -10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -74,107 +54,132 @@ class ExpandFromEdgeTests: XCTestCase {
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testExpandFromBottomEdgeOf() {
+    func testFlushWithTopEdgeOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-
-        let constraints = Constraid.expand(viewOne, fromBottomEdgeOf: viewTwo, times: 2.0, offsetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        let constraints = Constraid.flush(viewOne, withTopEdgeOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
         constraints.activate()
+
+        let constraintOne = viewOne.constraints.first!
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.top)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.top)
+        XCTAssertEqual(constraintOne.constant, 10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
+
+        XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
+    }
+
+    func testFlushWithBottomEdgeOf() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+        let constraints = Constraid.flush(viewOne, withBottomEdgeOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        constraints.activate()
+
         let constraintOne = viewOne.constraints.first!
 
         XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.greaterThanOrEqual)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintOne.constant, 10.0)
-        XCTAssertEqual(constraintOne.multiplier, 2.0)
-        XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
-
-        XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
-    }
-
-    func testExpandFromHorizontalEdgesOf() {
-        let viewOne = UIView()
-        let viewTwo = UIView()
-
-        viewOne.addSubview(viewTwo)
-
-        let constraints = Constraid.expand(viewOne, fromHorizontalEdgesOf: viewTwo, times: 2.0, offsetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
-        constraints.activate()
-        let constraintOne = viewOne.constraints.first!
-        let constraintTwo = viewOne.constraints.last!
-
-        XCTAssertEqual(constraints, viewOne.constraints)
-        XCTAssertEqual(constraintOne.isActive, true)
-        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
-        XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.top)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.lessThanOrEqual)
-        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.top)
         XCTAssertEqual(constraintOne.constant, -10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
 
-        XCTAssertEqual(constraintTwo.isActive, true)
-        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
-        XCTAssertEqual(constraintTwo.firstAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintTwo.relation, LayoutRelation.greaterThanOrEqual)
-        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintTwo.constant, 10.0)
-        XCTAssertEqual(constraintTwo.multiplier, 2.0)
-        XCTAssertEqual(constraintTwo.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
-
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testExpandFromVerticalEdgesOf() {
+    func testFlushWithVerticalEdgesOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-
-        let constraints = Constraid.expand(viewOne, fromVerticalEdgesOf: viewTwo, times: 2.0, offsetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        let constraints = Constraid.flush(viewOne, withVerticalEdgesOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
         constraints.activate()
+
         let constraintOne = viewOne.constraints.first!
         let constraintTwo = viewOne.constraints.last!
 
         XCTAssertEqual(constraints, viewOne.constraints)
+
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.lessThanOrEqual)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.constant, 10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
 
         XCTAssertEqual(constraintTwo.isActive, true)
         XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintTwo.firstAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintTwo.relation, LayoutRelation.greaterThanOrEqual)
+        XCTAssertEqual(constraintTwo.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintTwo.constant, 10.0)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
         XCTAssertEqual(constraintTwo.multiplier, 2.0)
         XCTAssertEqual(constraintTwo.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
 
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testExpandFromEdgesOf() {
+    func testFlushWithHorizontalEdgesOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
+        let constraints = Constraid.flush(viewOne, withHorizontalEdgesOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        constraints.activate()
 
-        let constraints = Constraid.expand(viewOne, fromEdgesOf: viewTwo, times: 2.0, offsetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        let constraintOne = viewOne.constraints.first!
+        let constraintTwo = viewOne.constraints.last!
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.top)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.top)
+        XCTAssertEqual(constraintOne.constant, 10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
+
+        XCTAssertEqual(constraintTwo.isActive, true)
+        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintTwo.firstAttribute, LayoutAttribute.bottom)
+        XCTAssertEqual(constraintTwo.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.bottom)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
+        XCTAssertEqual(constraintTwo.multiplier, 2.0)
+        XCTAssertEqual(constraintTwo.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
+
+        XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
+    }
+
+    func testFlushWithEdgesOf() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+        let constraints = Constraid.flush(viewOne, withEdgesOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
         constraints.activate()
 
         let constraintOne = viewOne.constraints[0]
@@ -187,43 +192,44 @@ class ExpandFromEdgeTests: XCTestCase {
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.top)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.lessThanOrEqual)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.top)
-        XCTAssertEqual(constraintOne.constant, -10.0)
+        XCTAssertEqual(constraintOne.constant, 10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
 
         XCTAssertEqual(constraintTwo.isActive, true)
         XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintTwo.firstAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintTwo.relation, LayoutRelation.greaterThanOrEqual)
+        XCTAssertEqual(constraintTwo.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintTwo.constant, 10.0)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
         XCTAssertEqual(constraintTwo.multiplier, 2.0)
         XCTAssertEqual(constraintTwo.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
 
         XCTAssertEqual(constraintThree.isActive, true)
         XCTAssertEqual(constraintThree.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintThree.firstAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintThree.relation, LayoutRelation.lessThanOrEqual)
+        XCTAssertEqual(constraintThree.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintThree.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintThree.secondAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintThree.constant, -10.0)
+        XCTAssertEqual(constraintThree.constant, 10.0)
         XCTAssertEqual(constraintThree.multiplier, 2.0)
         XCTAssertEqual(constraintThree.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
 
         XCTAssertEqual(constraintFour.isActive, true)
         XCTAssertEqual(constraintFour.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintFour.firstAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintFour.relation, LayoutRelation.greaterThanOrEqual)
+        XCTAssertEqual(constraintFour.relation, LayoutRelation.equal)
         XCTAssertEqual(constraintFour.secondItem as! UIView, viewTwo)
         XCTAssertEqual(constraintFour.secondAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintFour.constant, 10.0)
+        XCTAssertEqual(constraintFour.constant, -10.0)
         XCTAssertEqual(constraintFour.multiplier, 2.0)
         XCTAssertEqual(constraintFour.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
 
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 }
+#endif
