+++ Tests/ConstraidTests/LimitByMarginTests.swift
@@ -1,25 +1,27 @@
 import XCTest
 import Constraid
 
-class FlushWithEdgesTests: XCTestCase {
+#if os(iOS)
+import UIKit
 
-    func testFlushWithLeadingEdgeOf() {
+class LimitByMarginTests: XCTestCase {
+    func testLimitByLeadingMarginOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-        let constraints = Constraid.flush(viewOne, withLeadingEdgeOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
-        constraints.activate()
 
+        let constraints = Constraid.limit(viewOne, byLeadingMarginOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        constraints.activate()
         let constraintOne = viewOne.constraints.first!
 
         XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.greaterThanOrEqual)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.leading)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.leadingMargin)
         XCTAssertEqual(constraintOne.constant, 10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -27,23 +29,23 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testFlushWithTrailingEdgeOf() {
+    func testLimitByTrailingMarginOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-        let constraints = Constraid.flush(viewOne, withTrailingEdgeOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
-        constraints.activate()
 
+        let constraints = Constraid.limit(viewOne, byTrailingMarginOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        constraints.activate()
         let constraintOne = viewOne.constraints.first!
 
         XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.lessThanOrEqual)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.trailing)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.trailingMargin)
         XCTAssertEqual(constraintOne.constant, -10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -51,23 +53,23 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testFlushWithTopEdgeOf() {
+    func testLimitByTopMarginOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-        let constraints = Constraid.flush(viewOne, withTopEdgeOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
-        constraints.activate()
 
+        let constraints = Constraid.limit(viewOne, byTopMarginOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        constraints.activate()
         let constraintOne = viewOne.constraints.first!
 
         XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.top)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.greaterThanOrEqual)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.top)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.topMargin)
         XCTAssertEqual(constraintOne.constant, 10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -75,23 +77,23 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testFlushWithBottomEdgeOf() {
+    func testLimitByBottomMarginOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-        let constraints = Constraid.flush(viewOne, withBottomEdgeOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
-        constraints.activate()
 
+        let constraints = Constraid.limit(viewOne, byBottomMarginOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        constraints.activate()
         let constraintOne = viewOne.constraints.first!
 
         XCTAssertEqual(constraints, viewOne.constraints)
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.lessThanOrEqual)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.bottom)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.bottomMargin)
         XCTAssertEqual(constraintOne.constant, -10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -99,61 +101,24 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testFlushWithVerticalEdgesOf() {
+    func testLimitByHorizontalMarginsOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-        let constraints = Constraid.flush(viewOne, withVerticalEdgesOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
-        constraints.activate()
 
+        let constraints = Constraid.limit(viewOne, byHorizontalMarginsOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        constraints.activate()
         let constraintOne = viewOne.constraints.first!
         let constraintTwo = viewOne.constraints.last!
 
         XCTAssertEqual(constraints, viewOne.constraints)
-
-        XCTAssertEqual(constraintOne.isActive, true)
-        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
-        XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
-        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintOne.constant, 10.0)
-        XCTAssertEqual(constraintOne.multiplier, 2.0)
-        XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
-
-        XCTAssertEqual(constraintTwo.isActive, true)
-        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
-        XCTAssertEqual(constraintTwo.firstAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintTwo.relation, LayoutRelation.equal)
-        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintTwo.constant, -10.0)
-        XCTAssertEqual(constraintTwo.multiplier, 2.0)
-        XCTAssertEqual(constraintTwo.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
-
-        XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
-    }
-
-    func testFlushWithHorizontalEdgesOf() {
-        let viewOne = UIView()
-        let viewTwo = UIView()
-
-        viewOne.addSubview(viewTwo)
-        let constraints = Constraid.flush(viewOne, withHorizontalEdgesOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
-        constraints.activate()
-
-        let constraintOne = viewOne.constraints.first!
-        let constraintTwo = viewOne.constraints.last!
-
-        XCTAssertEqual(constraints, viewOne.constraints)
-
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.top)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.greaterThanOrEqual)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.top)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.topMargin)
         XCTAssertEqual(constraintOne.constant, 10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -161,9 +126,9 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(constraintTwo.isActive, true)
         XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintTwo.firstAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintTwo.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintTwo.relation, LayoutRelation.lessThanOrEqual)
         XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.bottom)
+        XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.bottomMargin)
         XCTAssertEqual(constraintTwo.constant, -10.0)
         XCTAssertEqual(constraintTwo.multiplier, 2.0)
         XCTAssertEqual(constraintTwo.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -171,12 +136,48 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 
-    func testFlushWithEdgesOf() {
+    func testLimitByVerticalMarginsOf() {
         let viewOne = UIView()
         let viewTwo = UIView()
 
         viewOne.addSubview(viewTwo)
-        let constraints = Constraid.flush(viewOne, withEdgesOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+
+        let constraints = Constraid.limit(viewOne, byVerticalMarginsOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
+        constraints.activate()
+        let constraintOne = viewOne.constraints.first!
+        let constraintTwo = viewOne.constraints.last!
+
+        XCTAssertEqual(constraints, viewOne.constraints)
+        XCTAssertEqual(constraintOne.isActive, true)
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.leading)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.greaterThanOrEqual)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.leadingMargin)
+        XCTAssertEqual(constraintOne.constant, 10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
+
+        XCTAssertEqual(constraintTwo.isActive, true)
+        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintTwo.firstAttribute, LayoutAttribute.trailing)
+        XCTAssertEqual(constraintTwo.relation, LayoutRelation.lessThanOrEqual)
+        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.trailingMargin)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
+        XCTAssertEqual(constraintTwo.multiplier, 2.0)
+        XCTAssertEqual(constraintTwo.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
+
+        XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
+    }
+
+    func testLimitByMarginsOf() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+
+        let constraints = Constraid.limit(viewOne, byMarginsOf: viewTwo, times: 2.0, insetBy: 10.0, priority: Constraid.LayoutPriority(rawValue: 500))
         constraints.activate()
 
         let constraintOne = viewOne.constraints[0]
@@ -189,9 +190,9 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(constraintOne.isActive, true)
         XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintOne.firstAttribute, LayoutAttribute.top)
-        XCTAssertEqual(constraintOne.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintOne.relation, LayoutRelation.greaterThanOrEqual)
         XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.top)
+        XCTAssertEqual(constraintOne.secondAttribute, LayoutAttribute.topMargin)
         XCTAssertEqual(constraintOne.constant, 10.0)
         XCTAssertEqual(constraintOne.multiplier, 2.0)
         XCTAssertEqual(constraintOne.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -199,9 +200,9 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(constraintTwo.isActive, true)
         XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintTwo.firstAttribute, LayoutAttribute.bottom)
-        XCTAssertEqual(constraintTwo.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintTwo.relation, LayoutRelation.lessThanOrEqual)
         XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.bottom)
+        XCTAssertEqual(constraintTwo.secondAttribute, LayoutAttribute.bottomMargin)
         XCTAssertEqual(constraintTwo.constant, -10.0)
         XCTAssertEqual(constraintTwo.multiplier, 2.0)
         XCTAssertEqual(constraintTwo.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -209,9 +210,9 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(constraintThree.isActive, true)
         XCTAssertEqual(constraintThree.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintThree.firstAttribute, LayoutAttribute.leading)
-        XCTAssertEqual(constraintThree.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintThree.relation, LayoutRelation.greaterThanOrEqual)
         XCTAssertEqual(constraintThree.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintThree.secondAttribute, LayoutAttribute.leading)
+        XCTAssertEqual(constraintThree.secondAttribute, LayoutAttribute.leadingMargin)
         XCTAssertEqual(constraintThree.constant, 10.0)
         XCTAssertEqual(constraintThree.multiplier, 2.0)
         XCTAssertEqual(constraintThree.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -219,9 +220,9 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(constraintFour.isActive, true)
         XCTAssertEqual(constraintFour.firstItem as! UIView, viewOne)
         XCTAssertEqual(constraintFour.firstAttribute, LayoutAttribute.trailing)
-        XCTAssertEqual(constraintFour.relation, LayoutRelation.equal)
+        XCTAssertEqual(constraintFour.relation, LayoutRelation.lessThanOrEqual)
         XCTAssertEqual(constraintFour.secondItem as! UIView, viewTwo)
-        XCTAssertEqual(constraintFour.secondAttribute, LayoutAttribute.trailing)
+        XCTAssertEqual(constraintFour.secondAttribute, LayoutAttribute.trailingMargin)
         XCTAssertEqual(constraintFour.constant, -10.0)
         XCTAssertEqual(constraintFour.multiplier, 2.0)
         XCTAssertEqual(constraintFour.priority, UILayoutPriority(rawValue: UILayoutPriority.RawValue(500)))
@@ -229,3 +230,4 @@ class FlushWithEdgesTests: XCTestCase {
         XCTAssertEqual(viewOne.translatesAutoresizingMaskIntoConstraints, false)
     }
 }
+#endif
